+++ src/journal/store/mod.rs
@@ -12,5 +12,42 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub(crate) mod compound;
 pub(crate) mod remote;
 pub(crate) mod segment;
+pub(crate) mod stream;
+
+pub(crate) use self::compound::CompoundSegmentReader;
+use self::remote::Client;
+pub use self::stream::{StreamReader, StreamWriter};
+use super::ReplicatePolicy;
+use crate::{storepb, Result};
+
+pub(crate) async fn build_compound_segment_reader(
+    policy: ReplicatePolicy,
+    stream_id: u64,
+    epoch: u32,
+    copy_set: Vec<String>,
+    start: Option<u32>,
+) -> Result<CompoundSegmentReader> {
+    // FIXME(w41ter) more efficient implementation.
+    let mut streamings = vec![];
+    for addr in copy_set {
+        let client = Client::connect(&addr).await?;
+        let req = storepb::ReadRequest {
+            stream_id,
+            seg_epoch: epoch,
+            start_index: start.unwrap_or(1),
+            include_pending_entries: false,
+            limit: 0,
+        };
+        streamings.push(client.read(req).await?);
+    }
+
+    Ok(CompoundSegmentReader::new(
+        policy,
+        epoch,
+        start.unwrap_or(1),
+        streamings,
+    ))
+}
